2021-11-04: Added WebSockets client Support
Added
- WebSockets client Support
- TLS Server Name Indication for STOMP
- 'obuspa -c event' signals a USP event (for testing purposes)

Fixed
- Compilation errors on GCC 11.2 cross compiler for ARM
- ControllerTrust parameter set permission was incorrectly additionally
considering the set permission of the parent object

+++ src/core/iso8601.c
@@ -174,7 +174,7 @@ iso8601_us_strftime(char *buf, size_t bufsiz, const struct timeval *tv)
 
         size_t sz;
         sz = strftime(buf, bufsiz-8, "%FT%T.", gmtime(&tv->tv_sec));
-        sz += sprintf(buf+strlen(buf), "%06ldZ", tv->tv_usec);
+        sz += sprintf(buf+strlen(buf), "%06ldZ", (long int)tv->tv_usec);
         return sz;
 }
